# all: reduce memory pressure, reduce allocations