--- a/plotter/testdata/polygon_holes_golden.eps
+++ b/plotter/testdata/polygon_holes_golden.eps
@@ -2,7 +2,7 @@
 %%Creator gonum.org/v1/plot/vg/vgeps
 %%Title: 
 %%BoundingBox: 0 0 100 100
-%%CreationDate: 2017-11-09 14:19:37.614165532 +1030 ACDT m=+0.187713815
+%%CreationDate: 2020-05-18 15:26:47.780474519 +0200 CEST m=+5.317328037
 %%Orientation: Portrait
 %%EndComments
 
@@ -93,19 +93,16 @@ stroke
 0 0 1 setrgbcolor
 newpath
 36.666 38.48 moveto
-36.666 38.48 lineto
 97.5 38.48 lineto
 97.5 79.677 lineto
 36.666 79.677 lineto
 closepath
 44.27 43.63 moveto
-44.27 43.63 lineto
 59.479 43.63 lineto
 59.479 53.929 lineto
 44.27 53.929 lineto
 closepath
 89.896 64.228 moveto
-89.896 64.228 lineto
 74.687 64.228 lineto
 74.687 74.527 lineto
 89.896 74.527 lineto