… plotter/testdata/polygon_holes_golden.svg