```
$> benchstat ref.txt new.txt
name    old time/op    new time/op    delta
Draw-4     777µs ± 2%     534µs ± 3%  -31.31%  (p=0.000 n=18+18)

name    old alloc/op   new alloc/op   delta
Draw-4     957kB ± 0%     495kB ± 0%  -48.28%  (p=0.000 n=16+19)

name    old allocs/op  new allocs/op  delta
Draw-4     1.16k ± 0%     0.60k ± 0%  -48.15%  (p=0.000 n=20+20)
```

Please take a look.

<!--
Checklist:

- API changes have been discussed
- code is goformated correctly (goimports)
- packages with generated code have had code generation run
- tests pass locally
- linked to relevant issues

Please make sure your commit message summary line and pull request
title match the Go convention; a one-line summary of the change,
prefixed by the primary affected package that should complete the
sentence, "This change modifies Gonum to _____."
-->