@sbinet 2020-05-16
this is probably hinting at the need of pools of `vg.PathComp`s and `vg.Point`s.

here is the cpu profile of the above benchmark:
```
$> go tool pprof cpu-1.prof 
File: tst-canvas.test
Type: cpu
Time: May 16, 2020 at 4:06pm (CEST)
Duration: 1.30s, Total samples = 1.48s (113.49%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top
Showing nodes accounting for 690ms, 46.62% of 1480ms total
Showing top 10 nodes out of 173
      flat  flat%   sum%        cum   cum%
     110ms  7.43%  7.43%      110ms  7.43%  runtime.memclrNoHeapPointers
      90ms  6.08% 13.51%      150ms 10.14%  gonum.org/v1/plot/vg/draw.clipPoly
      90ms  6.08% 19.59%       90ms  6.08%  runtime.(*pallocBits).summarize
      80ms  5.41% 25.00%      210ms 14.19%  gonum.org/v1/plot/vg/draw.clipLine
      70ms  4.73% 29.73%      140ms  9.46%  gonum.org/v1/plot/vg.(*Path).Line
      70ms  4.73% 34.46%      220ms 14.86%  runtime.scanobject
      60ms  4.05% 38.51%       60ms  4.05%  runtime.pageIndexOf (inline)
      40ms  2.70% 41.22%      100ms  6.76%  gonum.org/v1/plot.talbotLinHanrahan
      40ms  2.70% 43.92%       40ms  2.70%  runtime.duffcopy
      40ms  2.70% 46.62%       40ms  2.70%  runtime.futex
```
![profile001](https://user-images.githubusercontent.com/1070920/82121730-8672a800-978f-11ea-97e3-300e0978e0c5.png)